+++ earth2studio/models/px/pangu.py
@@ -24,7 +24,6 @@
 # nor does it submit to any jurisdiction.
 
 
-import os
 from collections import OrderedDict
 from collections.abc import Generator, Iterator
 from typing import TypeVar
@@ -43,6 +42,7 @@ from earth2studio.models.auto import AutoModelMixin, Package
 from earth2studio.models.batch import batch_func
 from earth2studio.models.px.base import PrognosticModel
 from earth2studio.models.px.utils import PrognosticMixin
+from earth2studio.models.utils import create_ort_session
 from earth2studio.utils import handshake_coords, handshake_dim
 from earth2studio.utils.type import CoordSystem
 
@@ -199,69 +199,9 @@ class PanguBase(torch.nn.Module, AutoModelMixin, PrognosticMixin):
             if self.ort is not None:
                 model_path = self.ort._model_path
                 del self.ort
-                self.ort = self.create_ort_session(model_path, device)
+                self.ort = create_ort_session(model_path, device)
 
         return self
-
-    @staticmethod
-    def create_ort_session(
-        onnx_file: str,
-        device: torch.device = torch.device("cpu", 0),
-    ) -> InferenceSession:
-        """Create ORT session on specified device
-
-        Parameters
-        ----------
-        onnx_file : str
-            ONNX file
-        device : torch.device, optional
-            Device for session to run on, by default "cpu"
-
-        Returns
-        -------
-        ort.InferenceSession
-            ORT inference session
-        """
-        if ort is None:
-            raise ImportError(
-                "onnxruntime (onnxruntime-gpu) is required for Pangu. See model install notes for details.\n"
-                + "https://nvidia.github.io/earth2studio/userguide/about/install.html#model-dependencies"
-            )
-        options = ort.SessionOptions()
-        options.enable_cpu_mem_arena = False
-        options.enable_mem_pattern = False
-        options.enable_mem_reuse = False
-        options.intra_op_num_threads = 1
-
-        # That will trigger a FileNotFoundError
-        os.stat(onnx_file)
-        if device.type == "cuda":
-            if device.index is None:
-                device_index = torch.cuda.current_device()
-            else:
-                device_index = device.index
-
-            providers = [
-                (
-                    "CUDAExecutionProvider",
-                    {
-                        "device_id": device_index,
-                    },
-                ),
-                "CPUExecutionProvider",
-            ]
-        else:
-            providers = [
-                "CPUExecutionProvider",
-            ]
-
-        ort_session = ort.InferenceSession(
-            onnx_file,
-            sess_options=options,
-            providers=providers,
-        )
-
-        return ort_session
 
     @torch.inference_mode()
     def _forward(
@@ -389,9 +329,7 @@ class Pangu24(PanguBase):
     ):
         super().__init__()
 
-        self.ort: ort.InferenceSession = PanguBase.create_ort_session(
-            ort_24hr, self.device
-        )
+        self.ort: ort.InferenceSession = create_ort_session(ort_24hr, self.device)
         self._output_coords["lead_time"] = np.array([np.timedelta64(24, "h")])
 
     @classmethod
@@ -500,9 +438,7 @@ class Pangu6(PanguBase):
     ):
         super().__init__()
         # Only require 6 hour to load session on construction
-        self.ort: ort.InferenceSession = PanguBase.create_ort_session(
-            ort_6hr, self.device
-        )
+        self.ort: ort.InferenceSession = create_ort_session(ort_6hr, self.device)
         self.ort24 = ort_24hr
         self._output_coords["lead_time"] = np.array([np.timedelta64(6, "h")])
 
@@ -553,7 +489,7 @@ class Pangu6(PanguBase):
         coords = coords.copy()
 
         # Load other sessions (note .to() does not impact these)
-        ort24 = PanguBase.create_ort_session(self.ort24, self.device)
+        ort24 = create_ort_session(self.ort24, self.device)
 
         for i, (key, value) in enumerate(self.input_coords.items()):
             if key != "batch":
@@ -625,9 +561,7 @@ class Pangu3(PanguBase):
     ):
         super().__init__()
         # Only require 3 hour to load session on construction
-        self.ort: ort.InferenceSession = PanguBase.create_ort_session(
-            ort_3hr, self.device
-        )
+        self.ort: ort.InferenceSession = create_ort_session(ort_3hr, self.device)
         self.ort24 = ort_24hr
         self.ort6 = ort_6hr
         self._output_coords["lead_time"] = np.array([np.timedelta64(3, "h")])
@@ -680,8 +614,8 @@ class Pangu3(PanguBase):
         coords = coords.copy()
 
         # Load other sessions (note that .to() does not impact these)
-        ort24 = PanguBase.create_ort_session(self.ort24, self.device)
-        ort6 = PanguBase.create_ort_session(self.ort6, self.device)
+        ort24 = create_ort_session(self.ort24, self.device)
+        ort6 = create_ort_session(self.ort6, self.device)
 
         for i, (key, value) in enumerate(self.input_coords.items()):
             if key != "batch":
